Moved tilde expansion unit test to utility module
This should have also moved when 'tilde_expansion()' was moved here.
